Move job status module (#8856)
* Move JobStatus so that other modules can use it

* Address rubocop complaintes

* Improve the namespacing for the JobTracker
